增加字符串表格类

+++ src/common/time_tools.cpp
@@ -12,7 +12,7 @@
 
 namespace ctm
 {
-	unsigned long long MilliTimestamp()
+	unsigned long MilliTimestamp()
 	{
 #ifdef WIN32
 		SYSTEMTIME sys_time;
@@ -120,16 +120,16 @@ namespace ctm
 		return S2D(buf);
 	}
 
-	string MilliTimestamp2DateTime(unsigned long long time)
+	string MilliTimestamp2DateTime(unsigned long time)
 	{
 		char buf[64] = {0};
-		snprintf(buf, sizeof(buf), "%s.%03d", Timestamp2FormatDateTime(time / 1000).c_str(), (int)time % 1000);
+		snprintf(buf, sizeof(buf), "%s.%03u", Timestamp2FormatDateTime(time / 1000).c_str(), (unsigned int)time % 1000);
 		return string(buf);
 	}
 
 	string CClock::RunInfo() const
 	{
-		unsigned long long currentTime = MilliTimestamp();
+		unsigned long currentTime = MilliTimestamp();
 		string head = "[" + m_tips + "]";
 		string info = head + "Begin time : " + MilliTimestamp2DateTime(m_begin);
 		info += "\n" + head + "End time : " + MilliTimestamp2DateTime(currentTime);
